feat: separate pages (add main & not found page)
