+++ src/App.tsx
@@ -1,56 +1,16 @@
-import { useEffect, useState } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
+import { BrowserRouter, Route, Routes } from 'react-router-dom'
+import MainPage from './pages/Main'
+import NotFoundPage from './pages/NotFound'
 import './App.css'
-import { UserAPI } from './service/api/users'
-import { User } from './common/models/user'
 
 const App = () => {
-  const [count, setCount] = useState(0)
-  const [users, setUsers] = useState<User[]>([])
-
-  const fetchUsers = async () => {
-    try {
-      const { data } = await UserAPI.getUsers()
-      setUsers(data)
-    } catch (error) {
-      console.error('Failed to fetch users', error)
-    }
-  }
-
-  useEffect(() => {
-    fetchUsers()
-  }, [])
-
   return (
-    <div className="App">
-      <div>
-        <a href="https://vitejs.dev" target="_blank" rel="noreferrer">
-          <img src={viteLogo} className="logo" alt="Vite logo" />
-        </a>
-        <a href="https://reactjs.org" target="_blank" rel="noreferrer">
-          <img src={reactLogo} className="logo react" alt="React logo" />
-        </a>
-      </div>
-      <h1>Vite + React</h1>
-      <div className="card">
-        <button onClick={() => setCount((count) => count + 1)}>
-          count is {count}
-        </button>
-        <p>
-          Edit <code>src/App.tsx</code> and save to test HMR
-        </p>
-      </div>
-      <p className="read-the-docs">
-        Click on the Vite and React logos to learn more
-      </p>
-
-      <div>
-        {users.map((user) => (
-          <div key={user.id}>{user.lastName}</div>
-        ))}
-      </div>
-    </div>
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<MainPage />} />
+        <Route path="*" element={<NotFoundPage />} />
+      </Routes>
+    </BrowserRouter>
   )
 }
