+++ src/pages/NotFound/index.tsx
@@ -1,0 +1,5 @@
+const NotFoundPage = () => {
+  return <div>Not Found Page</div>
+}
+
+export default NotFoundPage
